1、研发评估工作量变成double类型；2、用户类型代码调整；3、用户编辑界面bug修改

+++ MyTeam/Models/ReqManage/OutPoolTableResultExcel.cs
@@ -27,7 +27,7 @@ namespace MyTeam.Models
         public string ReqDesc { get; set; }
 
         [Display(Name = "研发评估工作量（人天）")]
-        public int? DevWorkload { get; set; }
+        public double? DevWorkload { get; set; }
 
         [Display(Name = "研发联系人/联系电话")]
         public string ReqDevPerson { get; set; }
