+++ MyTeam/Models/ReqManage/Req.cs
@@ -85,7 +85,7 @@ namespace MyTeam.Models
         public int ReqType { get; set; }
 
         [Display(Name = "研发评估工作量")]
-        public int? DevWorkload { get; set; }
+        public double? DevWorkload { get; set; }
 
         [Display(Name = "需求状态")]
         public int ReqStat { get; set; }
